src/orchestrator.py
"""Orchestrator - Coordinates the multi-agent analysis pipeline.

Routes tasks to appropriate Claude models:
  - Haiku: Fast screening, data parsing, simple lookups
  - Sonnet: Core technical/fundamental analysis
  - Opus: Deep reasoning, thesis validation, final synthesis

Uses Evidence Scorecard + Checklist Method (no fake percentages).
"""

from pathlib import Path
from typing import Optional

import yaml


class Orchestrator:
    """Main orchestrator for the trading analysis pipeline."""

    def __init__(self, config_path: str = "config/config.yaml"):
        self.config = self._load_config(config_path)

    def _load_config(self, config_path: str) -> dict:
        path = Path(config_path)
        if not path.exists():
            raise FileNotFoundError(f"Config file not found: {config_path}")
        with open(path) as f:
            return yaml.safe_load(f)

    def run(
        self,
        symbol: str,
        asset_class: str = "equities",
        timeframe: str = "1d",
        thesis: Optional[str] = None,
        output_format: str = "markdown",
    ) -> str:
        """Execute the full analysis pipeline.

        Pipeline stages:
        1. Data Collection (Haiku) - Gather price data, fundamentals, news
        2. Technical Analysis (Sonnet) - Chart patterns, indicators, levels
        3. Fundamental Analysis (Sonnet) - Financials, valuation, catalysts
        4. Evidence Scorecard (Opus) - Compile supporting/contradicting evidence
        5. Checklist Validation (Opus) - Validate thesis against checklist
        6. Synthesis (Opus) - Final report with actionable insights
        """
        # TODO: Implement pipeline stages
        return f"[Placeholder] Analysis for {symbol} ({asset_class}, {timeframe})"

    def _collect_data(self, symbol: str, asset_class: str, timeframe: str) -> dict:
        """Stage 1: Collect raw data using Haiku for fast parsing."""
        # TODO: Implement data collection
        return {}

    def _analyze_technical(self, data: dict) -> dict:
        """Stage 2: Technical analysis using Sonnet."""
        # TODO: Implement technical analysis
        return {}

    def _analyze_fundamental(self, data: dict) -> dict:
        """Stage 3: Fundamental analysis using Sonnet."""
        # TODO: Implement fundamental analysis
        return {}

    def _build_scorecard(
        self, technical: dict, fundamental: dict, thesis: Optional[str]
    ) -> dict:
        """Stage 4: Build evidence scorecard using Opus.

        Evidence Scorecard format:
        - Supporting Evidence: List of concrete data points that support the thesis
        - Contradicting Evidence: List of concrete data points against the thesis
        - Neutral/Uncertain: Data points that could go either way
        - NO percentage scores or fake confidence numbers
        """
        # TODO: Implement evidence scorecard
        return {}

    def _validate_checklist(self, scorecard: dict) -> dict:
        """Stage 5: Run checklist validation using Opus.

        Checklist items are binary (PASS/FAIL/SKIP):
        - Is the trend aligned with the thesis?
        - Are key support/resistance levels respected?
        - Is volume confirming the move?
        - Are fundamentals supportive?
        - Is risk/reward acceptable?
        """
        # TODO: Implement checklist validation
        return {}

    def _synthesize(
        self,
        scorecard: dict,
        checklist: dict,
        output_format: str,
    ) -> str:
        """Stage 6: Generate final synthesis report using Opus."""
        # TODO: Implement synthesis
        return ""
