requirements.txt
# AI / LLM
anthropic>=0.39.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Configuration
pyyaml>=6.0.1

# HTTP / API clients
requests>=2.31.0
aiohttp>=3.9.0

# CLI
argparse
rich>=13.7.0

# Data sources
yfinance>=0.2.36

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Type checking
mypy>=1.8.0
